Cambios en el application.properties para la base de datos y corrección de errores
Se ha añadido en el application properties los comandos necesarios para el uso de la base de datos mysql, y se han corregido algunos  errores leves en la declaración de los repositorios.

+++ CheatHub/src/main/java/repositories/repositorioComentario.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import entities.Comentario;
 import entities.Publicacion;
 
-public interface repositorioComentario extends JpaRepository<Comentario,Long>{
+public interface repositorioComentario extends JpaRepository<Comentario,Integer>{
 	List<Comentario> findByNombreComentario(Publicacion publicacion);
 
 }
